requirements.txt
pandas==2.2.1
numpy==1.26.4
psycopg2-binary==2.9.9
